need to fix State does not extist on type Controller.Controller

+++ internal/Utils/MathUtils/MathUtils.go
@@ -25,7 +25,7 @@ func Clamp(v, max, min float64) float64 {
 	return v
 }
 
-func MapRange(num, inLow, inHigh, outLow, outHigh float32) float32 {
+func MapRange(num, inLow, inHigh, outLow, outHigh float64) float64 {
 	newNum := (outLow + (((num - inLow) / (inHigh - inLow)) * (outHigh - outLow)))
 	return newNum
 }
